requirements.txt
boto3==1.9.246
botocore==1.12.246
coloredlogs==14.0
docutils==0.15
emoji==0.5.4
docformatter==1.3.0
flake8-bugbear==19.8.0
flake8==3.7.8
fvcore==0.1.1.post20200716
gitdb2==2.0.5
GitPython==3.0.3
hydra-core==1.0.3
torch==1.4.0
joblib==0.14.1
nltk==3.4.5
numpy==1.17.5
omegaconf==2.0.2
pandas==1.1.1
pytest==5.3.2
pexpect==4.7.0
Pillow==7.2.0
py-gfm==0.1.4
py-rouge==1.1
pyyaml==5.3.1
pyzmq==18.1.0
regex==2020.1.8
myst-parser==0.12.2
attrs~=19.3
requests-mock==1.7.0
requests<3,>=2.21.0
scikit-learn==0.23.1
scipy==1.4.1
sh==1.12.14
sphinx_rtd_theme==0.4.3
sphinx-autodoc-typehints==1.10.3
Sphinx==2.2.0
subword-nmt==0.3.7
tensorboard==2.3.0
tensorboardX==2.1
tokenizers>=0.8.0
torchtext==0.6.0
tornado==6.0.4
tqdm==4.36.1
typing-extensions==3.7.4.1
Unidecode==1.1.1
websocket-client==0.56.0
websocket-server==0.4
jsonlines==1.2.0
transformers==3.5.1
Flask==1.1.2
rdflib==5.0.0
lxml==4.6.2
gensim==4.0.0b0
matplotlib==3.3.3
stanford-openie==1.0.1
datetime2==0.7.5
spacy==2.3.5
re101==0.4.0
websocket-client==0.56.0
websocket-server==0.4
bs4==0.0.1
tabulate==0.8.7
